pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "offlickr"
version = "0.1.0"
description = "Turn a Flickr 'Your Flickr Data' export into a self-contained offline website."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Yaniv Golan" }]
keywords = ["flickr", "archive", "photography", "static-site", "offline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: System :: Archiving",
]
dependencies = [
    "click>=8.1",
    "pydantic>=2.7",
    "nh3>=0.2.18",
    "python-slugify[unidecode]>=8.0",
    "rich>=13.7",
    "pillow>=10.3",
    "jinja2>=3.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
heic = ["pillow-heif>=0.16"]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "respx>=0.21",
    "freezegun>=1.5",
    "beautifulsoup4>=4.12",
    "syrupy>=4.6",
    "diff-cover>=9",
    "ruff>=0.5",
    "mypy>=1.10",
    "commitizen>=3.28",
    "pip-audit>=2.7",
]

[project.scripts]
offlickr = "offlickr.cli:main"

[project.urls]
Homepage = "https://github.com/yaniv-golan/offlickr"
Issues = "https://github.com/yaniv-golan/offlickr/issues"
Changelog = "https://github.com/yaniv-golan/offlickr/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/offlickr"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL", "RUF"]
ignore = ["PLR0913", "PLR2004"]  # PLR0913: too many args (CLI); PLR2004: magic values in tests

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
strict = true
python_version = "3.11"
files = ["src", "tests"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["-ra", "--strict-markers", "--strict-config"]
testpaths = ["tests"]
markers = [
    "perf: performance smoke tests (deselect with '-m \"not perf\"')",
]

[tool.coverage.run]
source = ["src/offlickr"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 75
exclude_also = [
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = ["pyproject.toml:version"]
